Remove `BufferThread` in favor of a refactored `SmartBufferThread`
`SmartBufferThread` can now be instantiated with the `PRESERVE = true`
template parameter to keep an internal reference to every allocated
smart pointer. This makes it so we kan keep the underlying memory alive
when it's passed around as raw pointers. We need this to emulate a
digitizer in the mocked ADQAPI.

This allows us to remove the `BufferThread` class, whose only purpose
was exactly that. We now use smart pointers everywhere.

Also remove the allocation capping mechanism (actually needed a mutex).
We'll restore it we need to.